--- tools/reviewer/web/src/app/dashboard/file-changes/code-block/add-comment-buttons/add-comment-buttons.component.html
@@ -1,17 +1,0 @@
-<div class="add-comment-buttons">
-  <div class="add-comment-button
-    {{ line.hasPlaceholder ? '' : 'closed-comment' }}
-    {{ codeBlockViewService.lineBackground(line, isNewCode) }}"
-    *ngFor="let line of lines; let i=index "
-    (mouseover)="fileChangesService.mouseHover(i, isNewCode)"
-    (click)="fileChangesService.openComments(i)"
-  >
-    <div class="icon">
-      <mat-icon
-        *ngIf="fileChangesService.hoverCheck(i, isNewCode) &&
-               !line.hasPlaceholder"
-      >add_box</mat-icon>
-    </div>
-    <div class="placeholder" [style.height.px]="line.height"></div>
-  </div>
-</div>
