--- tools/reviewer/web/src/app/dashboard/file-changes/code-block/add-comment-buttons/add-comment-buttons.component.scss
@@ -1,19 +1,0 @@
-@import '../code-block-variables';
-
-.add-comment-buttons {
-  >.add-comment-button {
-    width: 25px;
-    >.icon {
-      height: $line-height;
-      mat-icon {
-        color: #036fe8;
-      }
-    }
-    >.placeholder {
-      background: $comments-background;
-    }
-  }
-  >.closed-comment {
-    cursor: pointer;
-  }
-}
