--- tools/reviewer/web/src/app/dashboard/file-changes/code-block/add-comment-buttons/index.ts
@@ -1,1 +1,0 @@
-export * from './add-comment-buttons.component';
