--- tools/reviewer/web/src/app/dashboard/file-changes/code-block/changes-highlighting/changes-highlighting.component.html
@@ -1,14 +1,0 @@
-<div class="changes-highlighting">
-  <div class="line"
-    *ngFor="let line of lines"
-  >
-    <div class="line-changes">
-      <pre
-        [class]=" isNewCode ? 'new-code' : 'old-code' "
-        *ngIf="line.isChanged"
-        [HTML]="line.code"
-      ></pre>
-    </div>
-    <div class="placeholder" [style.height.px]="line.height"></div>
-  </div>
-</div>
