--- tools/reviewer/web/src/app/dashboard/file-changes/code-block/changes-highlighting/changes-highlighting.component.scss
@@ -1,20 +1,0 @@
-@import '../code-block-variables';
-
-.changes-highlighting {
-  width: 100%;
-  position: absolute;
-  top: 0;
-  left: 0;
-  >.line {
-    >.line-changes {
-      height: $line-height;
-      pre {
-        user-select: none;
-        color: rgba(0,0,0,0);
-        height: $line-height;
-        font-size: $code-size;
-        width: 100%;
-      }
-    }
-  }
-}
