--- tools/reviewer/web/src/app/dashboard/file-changes/code-block/changes-highlighting/changes-highlighting.component.ts
@@ -1,14 +1,0 @@
-import { Component, Input } from '@angular/core';
-
-import { Line } from '../code-block.component';
-
-// The component implements highlighting of code changes
-@Component({
-  selector: 'changes-highlighting',
-  templateUrl: './changes-highlighting.component.html',
-  styleUrls: ['./changes-highlighting.component.scss'],
-})
-export class ChangesHighlightingComponent {
-  @Input() lines: Line[];
-  @Input() isNewCode: boolean;
-}
