--- tools/reviewer/web/src/app/dashboard/file-changes/code-block/changes-highlighting/index.ts
@@ -1,1 +1,0 @@
-export * from './changes-highlighting.component';
