--- tools/reviewer/web/src/app/dashboard/file-changes/code-block/line-numbers/index.ts
@@ -1,1 +1,0 @@
-export * from './line-numbers.component';
