--- tools/reviewer/web/src/app/dashboard/file-changes/code-block/line-numbers/line-numbers.component.html
@@ -1,9 +1,0 @@
-<div class="line-numbers">
-  <div class="line-number
-    {{ codeBlockViewService.lineBackground(line, isNewCode) }}"
-    *ngFor="let line of lines; let i=index "
-  >
-    <pre>{{i+1}}</pre>
-    <div class="placeholder" [style.height.px]="line.height"></div>
-  </div>
-</div>
