--- tools/reviewer/web/src/app/dashboard/file-changes/code-block/line-numbers/line-numbers.component.scss
@@ -1,20 +1,0 @@
-@import '../code-block-variables';
-
-.line-numbers {
-  >.line-number {
-    >pre {
-      user-select: none;
-      direction: rtl;
-      opacity: 0.7;
-      padding: 0 3px;
-      font-size: $code-size;
-    }
-    >.placeholder {
-      width: 100%;
-      background: $comments-background;
-    }
-  }
-  >.default {
-    background: #fafbfc;
-  }
-}
