+++ tools/reviewer/web/src/app/dashboard/file-changes/file-changes.service.ts
@@ -17,45 +17,117 @@ interface AddCommentResponse {
 // diff component (and children) and send to another.
 @Injectable()
 export class FileChangesService {
-  // A number of the line, where user cursor is hovering right now.
-  private hoveredLine: number;
-  // Is cursor hovering above new code?
-  private isHoveredNewCode: boolean;
+  getDiff(diffId: string): void {
+    this.firebaseService.getDiff(diffId).subscribe(diff => {
+      this.diff = diff;
+      this.localThreads = this.diff
+        .getThreadList()
+        .filter(v => v.getFile().getFilename() === this.file.getFilename());
 
-  // Subjects, which send data:
-  lineHeightChanges = new Subject<HeightResponse>();
-  openCommentsChanges = new Subject<number>();
-  closeCommentsChanges = new Subject<number>();
-  addCommentChanges = new Subject<AddCommentResponse>();
-  deleteCommentChanges = new Subject<boolean>();
+      this.getBranchInfo();
+    });
+  }
 
-  // Methods, which receive data:
-  setLineHeight(param: HeightResponse): void {
-    this.lineHeightChanges.next(param);
+  getBranchInfo(): void {
+    this.localserverService
+      .getBranchInfo(this.diff.getId(), this.diff.getWorkspace())
+      .subscribe(branchInfo => {
+        this.branchInfo = branchInfo;
+        this.file = this.getFile(this.file.getFilename(), this.branchInfo);
+        this.getFileChanges(this.file);
+      });
   }
-  openComments(lineNumber: number): void {
-    this.openCommentsChanges.next(lineNumber);
+
+  // Get file from branchInfo by the filename
+  getFile(filename: string, branchInfo: BranchInfo): File {
+    const files: File[] = this.localserverService
+      .getFilesFromBranchInfo(branchInfo);
+    for (const file of files) {
+      if (filename === file.getFilename()) {
+        return file;
+      }
+    }
   }
-  closeComments(lineNumber: number): void {
-    this.closeCommentsChanges.next(lineNumber);
+
+  getFileChanges(currentFile: File): void {
+    // Left commit id - commit before the changes.
+    const leftCommitId: string = this.branchInfo
+      .getCommitList()[0]
+      .getId();
+
+    const leftFile: File = new File();
+    leftFile.setCommitId(leftCommitId);
+    leftFile.setFilename(currentFile.getFilename());
+    leftFile.setRepoId(this.branchInfo.getRepoId());
+    leftFile.setWorkspace(this.diff.getWorkspace());
+    const rightFile: File = currentFile;
+
+    this.localserverService
+      .getFileChanges(leftFile, rightFile)
+      .subscribe(textDiffResponse => {
+        this.textDiff = textDiffResponse.getTextDiff();
+
+        // TODO: use changes from localserver instead
+        this.changes = this.differenceService.compare(
+          this.textDiff.getLeftFileContents(),
+          this.textDiff.getRightFileContents(),
+        );
+
+        this.isLoading = false;
+      });
   }
-  addComment(param: AddCommentResponse): void {
-    this.addCommentChanges.next(param);
+
+  // Send diff with new comment to firebase
+  addComment(lineNumber: number, comments: Comment[]): void {
+    if (!this.file.getCommitId()) {
+      // TODO: Add more UX behavior here
+      // For example:
+      // Remove button 'add comment' if it's an uncommitted file.
+      // Or open uncommitted files in a special readonly mode.
+      // etc
+      this.notificationService
+        .error('Comment cannot be added to an uncommitted file');
+      return;
+    }
+
+    if (comments.length === 1) {
+      // Create new thread
+      const newThread: Thread = this.createNewThread(lineNumber, comments);
+      this.diff.addThread(newThread);
+    }
+
+    this.firebaseService.updateDiff(this.diff).subscribe(() => {
+      this.notificationService.success('Comment is saved in firebase');
+    });
   }
+
+  createNewThread(lineNumber: number, comments: Comment[]): Thread {
+    const newThread: Thread = new Thread();
+    newThread.setLineNumber(lineNumber);
+    newThread.setIsDone(false);
+    newThread.setCommentList(comments);
+    newThread.setRepoId(this.branchInfo.getRepoId());
+    newThread.setCommitId(this.file.getCommitId());
+    newThread.setFile(this.file);
+
+    return newThread;
+  }
+
   deleteComment(isDeleteThread: boolean): void {
-    this.deleteCommentChanges.next(isDeleteThread);
-  }
+    if (isDeleteThread) {
+      // Delete all threads without comments.
+      const threads: Thread[] = this.diff.getThreadList();
+      threads.forEach((thread, threadIndex) => {
+        if (thread.getCommentList().length === 0) {
+          threads.splice(threadIndex, 1);
+        }
+      });
+      this.diff.setThreadList(threads);
+    }
 
-  // Lines detect mouse hover by the method
-  mouseHover(i: number, isNewCode: boolean): void {
-    this.hoveredLine = i;
-    this.isHoveredNewCode = isNewCode;
-  }
-
-  // Lines use it to know, should they display
-  // 'add-comment-button` on the line or not
-  hoverCheck(i: number, isNewCode: boolean): boolean {
-    return this.hoveredLine === i && this.isHoveredNewCode === isNewCode;
+    this.firebaseService.updateDiff(this.diff).subscribe(() => {
+      this.notificationService.success('Comment is deleted');
+    });
   }
 
   // Get langulage from filename. Example:
